api.models: collect all server-side models in a single file
api/models.py now contains all the server-side pydantic model
definitions. The common server and client-facing models are in the
analogous models.py file in kernelci-core.
