+++ api/models.py
@@ -7,10 +7,19 @@
 # at the moment
 # pylint: disable=too-few-public-methods
 
-"""User model definitions"""
+# pylint: disable=no-name-in-module
 
-from typing import Optional
-from pydantic import conlist, Field
+"""Server-side model definitions"""
+
+from datetime import datetime
+from typing import Optional, TypeVar
+from pydantic import (
+    BaseModel,
+    conlist,
+    Field,
+)
+from fastapi import Query
+from fastapi_pagination import LimitOffsetPage, LimitOffsetParams
 from fastapi_users.db import BeanieBaseUser
 from fastapi_users import schemas
 from beanie import (
@@ -18,8 +27,37 @@ from beanie import (
     Document,
     PydanticObjectId,
 )
+from bson import ObjectId
 from kernelci.api.models_base import DatabaseModel, ModelId
 
+
+# PubSub model definitions
+
+class Subscription(BaseModel):
+    """Pub/Sub subscription object model"""
+    id: int = Field(
+        description='Subscription ID'
+    )
+    channel: str = Field(
+        description='Subscription channel name'
+    )
+    user: str = Field(
+        description=("Username of the user that created the "
+                     "subscription (owner)")
+    )
+
+
+class SubscriptionStats(Subscription):
+    """Pub/Sub subscription statistics object model"""
+    created: datetime = Field(
+        description='Timestamp of connection creation'
+    )
+    last_poll: Optional[datetime] = Field(
+        description='Timestamp when connection last polled for data'
+    )
+
+
+# User model definitions
 
 class UserGroup(DatabaseModel):
     """API model to group associated user accounts"""
@@ -69,3 +107,28 @@ class UserUpdate(schemas.BaseUserUpdate):
     """Schema for updating a user"""
     username: Optional[Indexed(str, unique=True)]
     groups: Optional[conlist(str, unique_items=True)]
+
+
+# Pagination models
+
+class CustomLimitOffsetParams(LimitOffsetParams):
+    """Model to set custom constraint on limit
+
+    The model is required to redefine limit parameter to remove the number
+    validation on maximum value"""
+
+    limit: int = Query(50, ge=1, description="Page size limit")
+
+
+class PageModel(LimitOffsetPage[TypeVar("T")]):
+    """Model for pagination
+
+    This model is required to serialize paginated model data response"""
+
+    __params_type__ = CustomLimitOffsetParams
+
+    class Config:
+        """Configuration attributes for PageNode"""
+        json_encoders = {
+            ObjectId: str,
+        }
